--- src/main/java/duke/Command.java
@@ -1,5 +1,0 @@
-package duke;
-
-public class Command {
-
-}
