+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,0 +1,20 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public abstract class AddTodoCommand extends Command {
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+
+    }
+
+}
