+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public abstract class Command {
+
+    public abstract boolean isExit();
+
+    public abstract void execute(TaskList taskList, Ui ui, Storage storage);
+
+}
