+++ src/main/java/duke/task/TodoTask.java
@@ -1,0 +1,19 @@
+package duke.task;
+
+import duke.task.Task;
+
+public class TodoTask extends Task {
+
+    public TodoTask(String name) {
+        super(name);
+    }
+
+    @Override
+    public String toString() {
+        if (isCompleted()) {
+            return String.format("[T][1] %s", getTaskName());
+        } else {
+            return String.format("[T][0] %s", getTaskName());
+        }
+    }
+}
